requirements.txt
streamlit 
torch==2.3.0 
#torchvision==0.18.0
#torchaudio==2.3.0
transformers 
librosa 
speechrecognition
pipwin
pyaudio
numpy=1.24.4
google-generativeai
langchain
langchain_community
langchain-google-genai
datasets
sentencepiece
-e .